requirements.txt
asgiref==3.8.1
asttokens==2.4.1
branca==0.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.7
django-cors-headers==4.3.1
django-countries==7.6.1
django-filter==24.2
django-import-export==3.3.8
django-leaflet==0.30.1
django-seed==0.3.1
djangorestframework==3.15.1
djangorestframework-gis==1.0
et-xmlfile==1.1.0
executing==2.0.1
Faker==24.9.0
folium==0.17.0
fonttools==4.53.1
geographiclib==2.0
gunicorn==22.0.0
idna==3.7
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lat-lon-parser==1.3.0
lazy_loader==0.4
llvmlite==0.43.0
Markdown==3.6
MarkupPy==1.14
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
model-bakery==1.18.2
msgpack==1.0.8
nest-asyncio==1.6.0
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
proglog==0.1.10
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg==3.1.18
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.2
scipy==1.14.0
setuptools==72.1.0
shapely==2.0.4
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tablib==3.5.0
threadpoolctl==3.5.0
toposort==1.10
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
websockets==12.0
whitenoise==6.7.0
xlrd==2.0.1
xlwt==1.3.0
xyzservices==2024.6.0
python-dotenv==1.0.1
django-redis==5.4.0
uvicorn==0.30.6
celery==5.3.0
redis==5.0.8
django-celery-results==2.5.1
gevent==24.2.1
